Add cross validation for ensembled models
This commit adds the ability to specify a seed for a specific ordering of file
indices and a different set of validation files for each model in an ensemble.
Since the validation sets are not longer completely congruent, we no longer
perform several validation tasks in the ensemble averaging job. We also compute
the ensemble thresholds by taking the average of thresholds from all the models.